requirements.txt
streamlit
PyMuPDF
python-docx
python-pptx
pandas
markdown
sentence-transformers
faiss-cpu
openai 